Switch back to using signed 64-bit score type, as per the original version of zpar. Overflow occurs with 32-bit ints. 32-bit floats don't suffer from quite the same amount of overflow.

+++ src/common/depparser/implementations/arceager/depparser_macros.h
@@ -24,7 +24,7 @@ const unsigned DEP_LABEL_COUNT=CDependencyLabel::MAX_COUNT;
 const unsigned DEP_LABEL_COUNT = 1;
 #endif
 
-typedef float SCORE_TYPE ;
+typedef int64_t SCORE_TYPE ;
 #include "action.h"
 
 // supertag
